+++ src/components/Hooks/useTasks.js
@@ -1,5 +1,6 @@
 import { useState } from 'react'
 import { db } from '../../firebase'
+import swal from 'sweetalert'
 
 export default function useTasks() {
   const [tasks, setTasks] = useState([])
@@ -9,7 +10,16 @@ export default function useTasks() {
   }
 
   function deleteTask(todo) {
-    db.collection('tasks').doc(todo.id).delete()
+    swal({
+      text: 'Are you sure you want to delete this task?',
+      icon: 'warning',
+      buttons: true,
+      dangerMode: true,
+    }).then((willDelete) => {
+      if (willDelete) {
+        db.collection('tasks').doc(todo.id).delete()
+      }
+    })
   }
 
   return { deleteTask, handleCheckbox, tasks, setTasks }
